phase1/requirements.txt
anthropic>=0.25.0
openai-whisper>=20231117
pyttsx3>=2.90
pyaudio>=0.2.13
keyboard>=0.13.5
numpy>=1.24.0
colorama>=0.4.6
